# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ConnectHub social platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Local account authentication
  - name: feed
    description: Personalized feed and discovery
  - name: posts
    description: Post creation and social actions
  - name: users
    description: Profiles and relationships
  - name: notifications
    description: User notifications
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get the current session
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Invalid registration
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out the current account
      responses:
        "204":
          description: Logged out
  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Get the personalized home feed
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostFeed"
  /discover:
    get:
      operationId: getDiscover
      tags: [feed]
      summary: Get trending posts and suggested people
      responses:
        "200":
          description: Discovery content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discover"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by name or username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 80
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get the current profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update the current profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a public profile
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by a user
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostFeed"
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Updated relationship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Updated relationship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Edit a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "204":
          description: Deleted post
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Updated like state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Updated like state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [posts]
      summary: Get comments on a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Post comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [posts]
      summary: Comment on a post
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete a comment
      parameters:
        - $ref: "#/components/parameters/CommentId"
      responses:
        "204":
          description: Deleted comment
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for the current user
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [posts]
      summary: Request a direct upload URL for a post image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Signed upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: integer
    CommentId:
      name: commentId
      in: path
      required: true
      schema:
        type: integer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegisterInput:
      type: object
      required: [name, username, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        username:
          type: string
          minLength: 3
          maxLength: 24
          pattern: "^[a-zA-Z0-9_]+$"
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Session:
      type: object
      required: [authenticated, user]
      properties:
        authenticated:
          type: boolean
        user:
          anyOf:
            - $ref: "#/components/schemas/UserProfile"
            - type: "null"
    UserSummary:
      type: object
      required: [id, name, username, avatarUrl, bio]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          required: [followersCount, followingCount, postsCount, isFollowing, joinedAt]
          properties:
            followersCount:
              type: integer
            followingCount:
              type: integer
            postsCount:
              type: integer
            isFollowing:
              type: boolean
            joinedAt:
              type: string
              format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        bio:
          type: string
          maxLength: 160
        avatarUrl:
          type: ["string", "null"]
    Post:
      type: object
      required: [id, author, content, imageUrl, likesCount, commentsCount, isLiked, createdAt, updatedAt]
      properties:
        id:
          type: integer
        author:
          $ref: "#/components/schemas/UserSummary"
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: ["string", "null"]
    PostUpdate:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
        imageUrl:
          type: ["string", "null"]
    PostFeed:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type: ["string", "null"]
    Comment:
      type: object
      required: [id, author, content, createdAt]
      properties:
        id:
          type: integer
        author:
          $ref: "#/components/schemas/UserSummary"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
    LikeResult:
      type: object
      required: [liked, likesCount]
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
    FollowResult:
      type: object
      required: [following, followersCount]
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
    Notification:
      type: object
      required: [id, type, message, actor, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [like, comment, follow]
        message:
          type: string
        actor:
          $ref: "#/components/schemas/UserSummary"
        postId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Discover:
      type: object
      required: [trending, suggestions]
      properties:
        trending:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
    UploadInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/gif]
    UploadResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string