# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ConnectHub social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSummary } from './userSummary';

export interface Comment {
  id: number;
  author: UserSummary;
  content: string;
  createdAt: Date;
}