# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ConnectHub social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}